Create SubsystemControlsConfig.java
The SubsystemsControlsConfig class stores all the controllers, buttons, and triggers we will be using when driving and operating the robot. Right now, it only has the stuff necessary to drive the robot + the operator joystick since we haven't decided on button mappings as other subsystems are still in progress. More will be added to this class as robot development progresses.