Changed the code in Drive so that Auto uses OPEN_LOOP. Change this back when we have finished PID tuning Drive.
Added timer in Autonomous to drive forward a certain amount of time.  We should instead use telemetry to drive a certain distance.  This works for now to demonstrate that Auto makes the robot move.
Current code drives forward 2 seconds, pauses for 1 second, drives backwards 2 seconds.

Uncommented all the odometry and gyro (pigeon) declaration and initialization code in prep for Auto.